Move `TabManagerNavDelegate` to its own file.
